chat-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
websockets==12.0
python-dotenv==1.0.0
numpy==1.26.3
pyaudio==0.2.14
motor==3.3.2
pymongo==4.6.1